+++ expenses/RELEASES.md
@@ -439,3 +439,11 @@ Based on:
 - Speakeasy CLI 1.44.2 (2.35.9) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.23.1] https://www.npmjs.com/package/@codat/sync-for-expenses/v/0.23.1 - expenses
+
+## 2023-06-08 00:23:55
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.45.0 (2.37.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.24.0] https://www.npmjs.com/package/@codat/sync-for-expenses/v/0.24.0 - expenses